goldfish: Fix /cache partition type.
Starting with kernel 3.10, yaffs2 partitions are no longer
supported. This patch changes the type of the cache partition
image to 'ext4' to allow booting a system based on such kernel,
for example aosp_x86_64-userdebug.

This works after the following patch was applied to the emulator:

    https://android-review.googlesource.com/#/c/98204/

Note that this works properly because the platform build system
will extract the value to know how to generate cache.img at build
time properly.

  See $AOSP/build/tools/releasetools/img_from_target_files

BUG=15144004

Change-Id: I6c123c9baf84d13507a44b8205ec79746fb2628f