Add fingerprint HAL into goldfish device
This is to provide minimal fingerprint support in SDK system
images. It provides functions to enroll and process fingerprint.

For fingerprint emulation to work, additionally, the emulator need
to implement the fingerprint module that gets user inputs (touching
fingerprint sensor and removing finger from it) and sends those
inputs to the guest system.

Change-Id: Ifc97b8a9bd5ffdcb465994f4a8f9004b4ce73f64