Fix one bug for glGetString unaccepted parameter input
In dEQP-GLES2.functional.negative_api.state.get_string test, if input
parameter is not an accepted value, glGetString should return invalid
value. But now glGetString do not check the input value, so in this
situation function just return NO ERROR. To fix this issue, we should
check the input value and set the error state if needed.

Change-Id: I75c121f26937ae94dbd6a5042e15bca80335d075
Signed-off-by: Lizhe Liu <lizhe.liu@intel.com>