Optimize glGetIntegerv to reduce the cost of translating
Hardware GLES emulation in the Android platform is implemented
with a mix of components. Translator module is one of the most
important components in emulator. It translate the function
calls into calls to the appropriate desktop APIs and cost lots
of reference to complete it. So optimization must be beneficial
to us. GL_MAX_CUBE_MAP_TEXTURE_SIZE, GL_MAX_RENDERBUFFER_SIZE
and GL_MAX_TEXTURE_SIZE are the rough estimate value of the
largest texture which the OPENGL can handle. All of them are
constant value. So there is no need for us to translate them
from guest os to host os every time. We can cache the value
after first time and reuse it next time.

Change-Id: I1a0352c40b667bafd6ec1fb66a8060d7fe4a415d
Signed-off-by: Lizhe Liu <lizhe.liu@intel.com>