Clean up minor opengl errors/warnings.
There are always disconcerting messages printed to logcat
whenever developing apps on the emulator.

glUtilsParamSize and other functions
often fail with "unknow param 0x00..."

These are missing GLenums and types that for
some reason are not in glUtils.cpp
in OpenglCodecCommon.

There also has been a long-standing warning
that eglSurfaceAttrib is not implemented.

It seems we don't need to implement that to get things done,
so we will give a no-op implementation that suppresses
the warning.

Change-Id: Ie44fbdd10422dd824c9c3c3ea70c386023663969
(cherry picked from commit 215bb77438519bcf1b952e4d5b621abb0bf3ed82)