Emulator: remove expensive glCheckFramebufferStatus
The additional calls of glCheckFramebufferStatus in glDraw*
functions have very negative impact on framerate.

More specifically, the glCheckFramebufferStatus will flush
the write buffer of qemu-pipe and then do a block read until
the host finishes all the gl calls before this call and send
back results. This turns out to be costly, probably partially
due to the relatively longer latency (than real GPU) we have
in the GPU emulation layer -- it would be nice to hammer
that latency down in the near future.

bug: 27418143

Change-Id: Ic6d9e739562c6bd0b688d62805fda4d469b98759