Exit when QemuPipe is corrupted by GL
Previously when QemuPipe is corrupted, it prints an error message, and
the application will continue to use it, which leads to random behaviour
in GL on the host.

This patch crashes the APP and pops up the crash dialog when the pipe is
(going to be) corrupted. It would happen if the user asks GL to write /
read a NULL pointer (i.e. passing NULL to glGenBuffers). It might also
happen for other unknown reasons.

Change-Id: I303772029bb72a845fa0d4c79b0d38dba1fca2c7
(cherry picked from commit 5bc2615d233a2a5de39f1033be4f063d0ae8f5e4)