Don't optimize out rcUpdateColorBuffer
bug:28406160

If we want to use rcUpdateColorBuffer to
synchronize on the host in order to remove
out of order webcam frames, we need to call it
if (cb->hostHandle), not just when
the width and height are nonzero.

Otherwise, we get deadlocks.

Change-Id: Ib98774b935e8786e307dd38711a0f2f9ffea440e
(cherry picked from commit 6298285500362b217b73b96def2ce207e206171d)