Remove obsolete 'qemud' service.
The qemud service was a small daemon used to implement
a multiplexing communication channel over a tty between
guest processes and the emulator. It has been replaced
by the QEMU pipe device since at least API level 14, and
is thus not needed anymore.

This patch removes it from the system build.

IMPORTANT NOTE:
    Many of the sources include <hardware/qemud.h>, currently
    provided by hardware/libhware/include/, but this header
    uses the pipe directly to connect to the emulator. Such
    code is thus unaffected by this patch.

Change-Id: I652ac746f1d17f8b2cced2782269013f768e3739