Max video resolutions consistent with supported
Four very similar tests in android.media.cts.MediaPlayerTest had been
failing.  It turns out that the camera reports certain available
resolutions, but a separate specification of the max width and height
is given in a certain XML file, and has values less than those of the
max reported resolution.  Updating that specification to the
width/height of the max reported resolution made the tests work.

This CL is the same as the mnc-emu-dev CL

https://googleplex-android-review.git.corp.google.com/#/c/1231696

Change-Id: I9057212b05a8da6b984d857f3045ec602bd167cd