Add required video frame size
b/30253054

The test class android.media.cts.MediaRecorderTest specifies a frame
size of 176x144.  EmulatedFakeCamera does not specify this as a
supported video size.  (In fact, it doesn't specify *any* supported
video sizes.)  This causes a number of failures.  Fix is to
add 176x144 to the list of supported preview sizes and video sizes.

Copied from this mnc-emu-dev CL:

https://googleplex-android-review.git.corp.google.com/#/c/1239780

Change-Id: I561ee428432e5e2bc25944838d91813e5ba2f39b