Keystore HAL for CTS test
CTS tests require a hardware keystore. This CL implements a simulated
hardware keystore.

The idea is to instead of using software key store on the guest, we
implement a virtual keystore device, and move the keystore work to the
host. This CL merely accepts keystore commands and passes them to the
host via qemu pipe. The real implementation is on the host, which is based
on OpenSSL and the current software keystore implementation.

Doc here:
https://docs.google.com/a/google.com/document/d/1k6bqetNptL4NLWiD-8VezMnSRwEhHE2o94KaqIRoEOY/edit?usp=sharing

BUG: 37640731

Test: run cts -m CtsKeystoreTestCases

Change-Id: I6f051068569d7b2189339978fb3d3d7603fa50b3