Restart camera preview on pixel format change
Make sure the camera preview is restarted if the preview pixel format
changes. Otherwise preview frames will be delivered with the wrong
format. In addition to this also make sure that the camera only cares
about the preview pixel format, not the video pixel format. The video
pixel format parameter key is not even exposed in the higher level APIs
and CTS tests do not seem to exercise this setting, it seems to be an
unused parameter.

BUG: 30700822
Test: ran camera CTS tests
Change-Id: Ife3ad70ab8fe2f1f834a057bdbf5558f684ec7df
(cherry picked from commit f4455dcc522ed67052e911e245eef7927e76b8f2)