Respond to camera auto-focus requests
Even though the emulated cameras do not support auto-focus they are
required to respond to auto-focus commands. They should immediately
signal a successful auto-focus complete event to the client. This change
implements this behavior by setting a flag that is inspected as long as
the worker thread that generates preview frames is running. This is safe
since the documentation mentions that the camera must be in preview mode
for the auto-focus request to work.

BUG: 30700822
Test: ran camera CTS tests
Change-Id: I2be24c734255bba2dcb55a73b84b49f7ee3e2232
(cherry picked from commit 01089051c75ff1e00476191614911f6046862fb3)