Forget camera preview window on preview disabled
The camera client does not informat the HAL that a preview window has
been destroyed unless the preview is running. This means that the
preview window can be destroyed and then later on when preview is
enabled the HAL will attempt to use a destroyed window. This fixes that
by setting the preview window pointer to NULL whenever previews stop.
The client will then set the preview window to use once it has enabled
preview again and that new window will be used.

BUG: 30034008

Test: ran camera CTS tests
Change-Id: I3b372d8e3d9732e87d05535867a1f67eba9e3170
(cherry picked from commit 1d812bda533463a7781589134cba834576ef9958)