Consolidate camera preview enabled checks
A previous patch introduced a boolean to keep track of whether the
camera preview was running or not. Such a mechanism already existed in
the preview window class. This removes the boolean in favor of using the
older mechanism instead.

Test: ran camera CTS tests
Change-Id: Id62003c9596edf53d702c080455cf940524d4877
(cherry picked from commit 804fb7575829203017638401c93c987c11b1e120)