Use correct pixel format for camera videos
A previous change removed code that would use the video pixel format if
video recording was enabled. This change was made because it didn't seem
to have any effect and it was causing problems because the code could
not reliably detect when recording was enabled. It turns out the video
pixel format is important and that the reason it was causing problems
was because the recording hint parameter was not properly reset and did
not cause preview restarts when needed. This change makes sure to use
the correct pixel format and correctly handles the recording hint
parameter.

Test: ran camera CTS tests
Change-Id: I7466a750f799c89a2933ccd712386fe6380c01d7
(cherry picked from commit 61cad039974014f0c6f29abe4b3519741c36074d)