Set correct EXIF picture size in camera pictures
The EXIF data created by the camera had pixel dimensions set to zero by
default. Newer versions of the camera app requires that these dimensions
are non-zero. This change updates the default values to match the
picture dimensions. In order to update a value this change also removes
an entry if it already exists and then recreates it.. This is to make it
easier to update tags without having to reallocate memory.

Test: ran camera CTS tests
Change-Id: If2d1b9714fe16d55f4d2a68e2cd6927d1d4866f6
(cherry picked from commit 7ef96014694a23e2d117544f2cca19a8b7ac0e01)