Fix format of camera preview FPS range parameter
The parameter for the preview FPS range contains two integers separated
by a comma. In the existing code the two integers were placed inside
parentheses which matches the list of support preview FPS ranges. It
seems however that this is not valid, the parentheses only exist in the
list of supported ranges so that treating it as a list becomes easier.
Newer versions of the Camera app relies on this parameter having the
correct format as does the getPreviewFpsRange method in the
CameraParameters class.

Test: ran camera CTS tests
Change-Id: If95fb0f2db2116a9c6452ac78da8f5fa077ee602
(cherry picked from commit 887d8c6b8ff6d037cc652fe59f4c160fa6de0d46)