Fix camera incorrectly using YU12 instead of YV12
The pixel formats provided to the emulated camera are translated from
the camera parameter string value to a fourcc code. This translation was
incorrect for YV12, instead the camera interpreted it as YU12. This
caused the U and V components to be switched, altering the color of
images received when requesting YV12. This change makes sure that the
correct pixel format is used.

BUG: 32489697
Test: ran camera CTS tests
Change-Id: I65c908873631d58248fea6abad440125a91ac850
(cherry picked from commit 13650c14d59c7bd0247bb26e7e2599832581cc2d)