Start video on request even without recording hint
The camera parameter API contains a recording hint that we used to
determine if video recording was to be enabled or not. Unfortunately not
all clients use the recording hint so we cannot rely solely on that.
Luckily the camera is explicitly told when recording should start. This
change uses that mechanism to determine if we need to restart frame
delivery with a new pixel format or frame size because they might have
changed to something different than what the preview is using.

BUG: 32771006
Test: ran camera CTS tests
Change-Id: I0e0aac4915aa97d0cba71a4b652e6e2a68da7b6e
(cherry picked from commit 7705b5d1c27b8c4bb4865b1defcb160b0e1eecef)