Set correct camera orientation for webcams
This fixes the camera orientation for web cameras. Previously the
orientation would be 270 degrees which means that if you look at the
adjusted preview frame it will be upside down. With this change the
orientation is 90 degrees which means that it will line up with the way
JPEG pictures are generated.

BUG: 33245832
Change-Id: I25b4acbb05be0ac84eae000b96cb6e193b260ffa
(cherry picked from commit 1cc193d879dea74095aaa0bc22ee925d8b772e36)