Prevent camera moving square from getting stuck
The square in the emulated camera could get stuck bouncing around along
one edge if the time between two frames took long enough that the next
frame didn't clear the edge entirely. The coordinate for the square
would then remain outside the view and simply be negated once again,
causing it to bounce back immediately. On the next frame it would then
be negated again and so on.

This fixes the issue by correctly updating the relative position on
every bounce to make sure it's on the inside of the edge after the
bounce.

BUG: 34252276
Change-Id: I967fbef70d667e9977aaab9a11d07d7d6e8cd161
(cherry picked from commit 29dd289d7593bb4446b9d13a175a9cb2cfece01b)