Camera: v2/v3: Fix deadlock in sensor / result thread communication
Sometimes, the HAL locks up in a situation where
android::Sensor::waitForNewFrame and android::Sensor::threadLoop are
both waiting on the other to signal; specifically on lines
Sensor.cpp:294 and Sensor.cpp:198.

This results in FastBasicsTest#testCamera2 fail at a ~8% rate in
presubmit testing (assuming this is the only source of failures).

After some staring at the code, the likely issue is that
waitForNewFrame only signals on mReadoutComplete when it didn't wait
for mReadoutAvailable.  This isn't actually foolproof, if the sensor
threadloop manages to execute twice (and ends up waiting for
mReadoutComplete) before the readout thread is woken up from its wait
on mReadoutAvailable.

So make the signaling unconditional in waitForNewFrame.

Additionally switch to "HAL_PIXEL_FORMAT_YCBCR_420_888" for NV12 stream
case.

Test: Lots of local iterations of CTS
   android.hardware.camera2.cts.FastBasicsTest
Bug: 36233279

Change-Id: Ia64af66d1c5d908b8c081d0bb00adce282226924