Make Emulator's Light sensor recognized
The framework didn't like the Light Sensor to include
SENSOR_FLAG_WAKE_UP as one of its flags. With this
flag, the system did not recognize the sensor.

After removing this flag, the system recognizes the
Light Sensor and accepts data from it.

Test: Verified the behavior using the SensorKinetics app
Verified on 'master' and 'oc-emu-dev' branches.

The full app name is SensorKinetics_v2.1.2_apkpure.com.apk

Bug: 36726002

Change-Id: Id87c5e73804f1519ce46c37249765ca0966a185a