Replace libcamera_client with a VNDK lib
libcamera_client is not a VNDK library and thus not available to vendors
in a Treblized system. This camera HAL uses the library due to
CameraMetadata and CameraParamters classes. They are now available via
a static VNDK library, android.hardware.camera.common@1.0-helper, which
is available to vendors. Thus switching to the lib.

Bug: 65148308
Test: Verified camera HAL binary, camera.ranchu.so, to make sure it has
      no dependecy on libcamera_client.so.
Change-Id: I31a093834d1bac62d66e93416d989527bedafcfa