camera: add rotating fake camera
Cts verifier needs to test the Rotation vector
sensor, and this camera setup is soley for that
test.

It places the dot image (from the test) on the virtual
ground, moves the camera away at certain distance,
and rotates the camera round x/y/z axis, based on the
rotating input from emulator console.

The rotating fake camera is only turned on by the
qemu.camera.fake.rotating=1.

notes: the camcorder profile is changed to have higher
width and height (to 1280 by 720), and the encoder
cap is also changed accordingly to handle 1280 by 720.

dependency: emulator side will need to implement
FakeRotatingCameraSensor pipe service to provide the rotating
angles to this camera.

Test: emulator -prop qemu.camera.fake.rotating=1
run the rotation vector test and it should pass.

BUG: 65366470
Change-Id: Iae1a7ef61d6f2e6cae8530a2440f311581d7f505