Add software-only "strongbox" KM4 to goldfish.
This configuration emulates devices with an additional keymaster HAL on a
separate "StrongBox" security chip. (It does not provide the security properties
of a physical StrongBox.) This will allow tests on the emulator to run through
KeyStore code paths that select the StrongBox Keymaster.

Move trustry KM to goldfish/keymaster/trusty.

Test: (manual) emulator boots, strongbox KM4 can start

Change-Id: I665fec3fb55432c2b55a07c14e9a870d9d4c67a2