Use softkeymaster in guest instead of connecting to host
We added host implementation of softkeymaster in emulator by
copying code in guest from API 26 in order to pass VTS.
However, the softkeymaster implementation in the host would no
longer be compatible with new API levels at some point.

To avoid future incompatiblity, we decided to switch to
softkeymaster in guest.

Change-Id: I0c7766282f3b7dace4ace5f446bebe3962138d61
Signed-off-by: Weilun Du <wdu@google.com>