Fixes to support WiFi on newer version of Android
Replace ip for creating namespaces with a new command named createns.
This is because ip required /var/run/netns to exists, this no longer
works because we can't symlink /var in the root directory. The new
createns instead places the file in a vendor location and also provides
a file with a PID of a process in the new network namespace. This gets
rid of the looping shell trick in init.wifi.sh. This change also
requires some updates to execns to work with createns. Some additional
fixes for compilation errors.

BUG: 74514143
Test: Compile emulator images and verify that WiFi works
Change-Id: I00dcee2ee025a44ab5cdcc883570a39235b9b6d8