Use hostapd without HIDL for virtual access points
Change the init file to use the new hostapd binary that does not have
HIDL control interface support. This way we can have both HIDL and
non-HIDL hostapd binaries at the same time. The non-HIDL hostapd is used
to provide a virtual acccesss point and the HIDL enabled hostapd is used
to provide WiFi hotspot functionality. Modify the manifest to make sure
that the hostapd HIDL for hotspots is included.

BUG: 74401469
Test: run cts -m CtsNetTestCases
Change-Id: I6ad2af0415c53e50050d7b91c1beb7dd45f4bbdc