Fix error code returned from sim close channel
When a SIM_CLOSE_CHANNEL request fails because the channel number is
invalid the emulator returns CME Error: 21 which indicates that the
channel index is invalid. The reference RIL treats any error as a
generic failure so this information is lost. There is a VTS test that
expects the RIL error code (which is different from the CME error code)
to indicate that the channel provided is an invalid argument.

Inspect the return code from the modem and if it's invalid index
translate it to an invalid arguments error. Otherwise return generic
failure.

BUG: 110212792
Test: run vts -m VtsHalRadioV1_0Target -t
VtsHalRadioV1_0Target#RadioHidlTest.iccCloseLogicalChannel

Merged-In: I2617458c67f7470fab12efa6068347cac01e2564
(cherry picked from commit fdde7ebf0e6f1c53f0fb13da8680281d85a6cb76)