Return result after RIL set auto network selection
Our RIL was only returning the result of set network selection automatic
if the SIM card is absent. On other types of failure or success nothing
was returned. This results in a timeout in the upper RIL layers because
they never get to know if the request was successful or not.

Return results for successful or failed requests based on the modem's
response.  This fixes tests that rely on the return code to know whether
the operation succeeded or not.

BUG: 110212792
Test: run vts -m VtsHalRadioV1_0Target -t
VtsHalRadioV1_0Target#RadioHidlTest.setNetworkSelectionModeAutomatic

Change-Id: I82569285fb4f5274d4be72802aad00cffd5a91ec