Add emulator audio policy config in /vendor not /system
audio_policy_configuration_generic.xml is pushed in /system as the default
configuration (used if the vendor does not provide one in /vendor or /odm)
and for the emulator.

As a result it can be parsed by 2.0 and 4.0 audio HALs. The format
is retro-compatible, but not forward-compatible.

Additionally, the generic config pushed in /system should never be used
as the vendor should always provide a configuration file describing its
target architecture. Thus it should not be part of the GSI.

Those files were originally pushed in /system as the emulator
target does not provide its own configuration file.
The emulator should provide config file instead in /vendor instead.

Bug: 78543061
Test: sdk_gphone_x86 and aosp_x86_64 emulator has audio
Test: adb shell cat /vendor/etc/audio_policy_configuration.xml | diff - \
          frameworks/av/services/audiopolicy/config/audio_policy_configuration_generic.xml
Change-Id: Ice43cb4670a82f4e60fcccc58a636d224f240b5a
Signed-off-by: Kevin Rocard <krocard@google.com>