Add script combining multiple images
Current android build process generates multiple *-qemu.img in which
they only have one partition. For future development, a single image
with multiple partitions is needed. This python script combines raw
images and output a image file with multiple partitions.

Additionally, a simple document is added which includes the functions,
prerequisites, usage and examples of configuration file.

This is the first version of this tool and further TODOs and code
refactoring are needed.

go/combine-img to get a quick guide and get started

Bug: 111330873
Test: $ python mk_combined_img.py --input image_config
      $ cgdisk $OUT/combined.img
      # check if the partition table is correct
      $ /path/to/emulator -acpi-config acpi.ini -show-kernel
      $ adb root; adb shell
      # check the booting process

Change-Id: I4a3c6223f5d5f8990d47658f71eff2536033b174