Retain GPS flags across multiple messages
This code sets flags to indicate what GPS values are present.
The current code sets the flags only according to the most
recent message. This change aggregates the flags, so if
different message types are received, the flags indicate all
the values that were received in the multiple messages.

This is needed because the Emulator now sends both altitude
and speed, which requires two types of messages.

This CL also removes some dead code.

Bug: 112259283
Test: The VtsHalGnssV1_0Target group now passes 6/6
Change-Id: Iecf710f02a2bac6740f5beff729d6da9224868bc